# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Fooseball api Documentation",
        "description": "The endpoints to connect to the fooseball api",
        "contact": {
            "name": "Van de Voorde Siebe",
            "email": "siebe.vandevoorde@student.odisee.be"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "api/tables/{tableId}/teams/{teamId}": {
            "patch": {
                "tags": [
                    "Gameinfo"
                ],
                "summary": "Update score",
                "description": "Update the specified resource in storage.",
                "operationId": "914ca45c4d078d9c5293957eedd90650",
                "parameters": [
                    {
                        "name": "tableId",
                        "in": "path",
                        "description": "ID of the foosball table",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "teamId",
                        "in": "path",
                        "description": "ID of the team",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Score change",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "score_change": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Score updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Score updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid score change."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No game is active on this table."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "api/tables/{id}/start": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Start game",
                "description": "Start a game.",
                "operationId": "f0a5ea2b33d07ea7327fbb39d1123e5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the foosball table",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game started successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Game started successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "208": {
                        "description": "Game is already running",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Game is already running."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Foosball table not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/tables/{id}/end": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "End game",
                "description": "Stop a game.",
                "operationId": "b28822fa0551a7e8f6d23f08e43446c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the foosball table",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game ended successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Game ended successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "208": {
                        "description": "No Game is running",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No Game is running."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Game not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}